feat(go): start adding series controllers

+++ kiwiscript_go/services/series.service.go
@@ -26,41 +26,57 @@ import (
 	"github.com/kiwiscript/kiwiscript_go/utils"
 )
 
-func (s *Services) FindSeriesBySlug(ctx context.Context, slug string) (db.Series, *ServiceError) {
-	log := s.log.WithGroup("service.series.FindSeriesBySlug").With("slug", slug)
-	log.InfoContext(ctx, "Getting series by slug")
-	series, err := s.database.FindSeriesBySlug(ctx, slug)
+type FindSeriesBySlugsOptions struct {
+	LanguageSlug string
+	SeriesSlug   string
+}
 
+func (s *Services) FindSeriesBySlugs(ctx context.Context, opts FindSeriesBySlugsOptions) (*db.Series, *ServiceError) {
+	log := s.
+		log.
+		WithGroup("service.series.FindSeriesBySlug").
+		With("laguageSlug", opts.LanguageSlug, "slug", opts.SeriesSlug)
+	log.InfoContext(ctx, "Getting series by slug")
+
+	language, serviceErr := s.FindLanguageBySlug(ctx, opts.LanguageSlug)
+	if serviceErr != nil {
+		log.WarnContext(ctx, "Language not found", "error", serviceErr)
+		return nil, serviceErr
+	}
+
+	series, err := s.database.FindSeriesBySlugAndLanguageID(ctx, db.FindSeriesBySlugAndLanguageIDParams{
+		Slug:       opts.SeriesSlug,
+		LanguageID: language.ID,
+	})
 	if err != nil {
 		log.Warn("Error getting series by slug", "error", err)
-		return series, FromDBError(err)
+		return nil, FromDBError(err)
 	}
 
 	log.InfoContext(ctx, "Series found")
-	return series, nil
+	return &series, nil
 }
 
 type findOrCreateTagOptions struct {
-	Length   int
 	Tags     []string
 	SeriesID int32
 	UserID   int32
 }
 
-func (s *Services) findOrCreateTags(ctx context.Context, qrs *db.Queries, options findOrCreateTagOptions) ([]db.Tag, *ServiceError) {
-	tags := make([]db.Tag, options.Length)
+func (s *Services) findOrCreateTags(ctx context.Context, qrs *db.Queries, opts findOrCreateTagOptions) ([]db.Tag, *ServiceError) {
+	tags := make([]db.Tag, len(opts.Tags))
 
-	for i, name := range options.Tags {
+	for i, name := range opts.Tags {
 		tag, err := qrs.FindOrCreateTag(ctx, db.FindOrCreateTagParams{
 			Name:     name,
-			AuthorID: options.UserID,
+			AuthorID: opts.UserID,
 		})
 		if err != nil {
 			return nil, FromDBError(err)
 		}
 
 		params := db.CreateSeriesTagParams{
-			SeriesID: options.SeriesID,
+			SeriesID: opts.SeriesID,
 			TagID:    tag.ID,
 		}
 		if err := qrs.CreateSeriesTag(ctx, params); err != nil {
@@ -74,70 +90,77 @@ func (s *Services) findOrCreateTags(ctx context.Context, qrs *db.Queries, option
 }
 
 type CreateSeriesOptions struct {
-	UserID      int32
-	Title       string
-	Description string
-	Tags        []string
+	UserID       int32
+	LanguageSlug string
+	Title        string
+	Description  string
+	Tags         []string
 }
 
-func (s *Services) CreateSeries(ctx context.Context, options CreateSeriesOptions) (db.Series, []db.Tag, *ServiceError) {
+func (s *Services) CreateSeries(ctx context.Context, options CreateSeriesOptions) (*db.Series, []db.Tag, *ServiceError) {
 	log := s.log.WithGroup("service.series.CreateSeries")
 	log.InfoContext(ctx, "create series", "title", options.Title)
-	slug := utils.Slugify(options.Title)
 
-	if _, serviceErr := s.FindSeriesBySlug(ctx, slug); serviceErr == nil {
+	language, serviceErr := s.FindLanguageBySlug(ctx, options.LanguageSlug)
+	if serviceErr != nil {
+		return nil, nil, NewError(CodeNotFound, MessageNotFound)
+	}
+
+	slug := utils.Slugify(options.Title)
+	params := db.FindSeriesBySlugAndLanguageIDParams{
+		Slug:       slug,
+		LanguageID: language.ID,
+	}
+	if _, err := s.database.FindSeriesBySlugAndLanguageID(ctx, params); err == nil {
 		log.InfoContext(ctx, "series already exists", "slug", slug)
-		return db.Series{}, nil, NewDuplicateKeyError("series already exists")
+		return nil, nil, NewDuplicateKeyError("Series already exists")
 
 	}
 
-	tLen := len(options.Tags)
-	if tLen == 0 {
+	if options.Tags == nil || len(options.Tags) == 0 {
 		log.InfoContext(ctx, "create series without tags")
 		series, err := s.database.CreateSeries(ctx, db.CreateSeriesParams{
 			Title:       options.Title,
 			Description: options.Description,
 			AuthorID:    options.UserID,
+			LanguageID:  language.ID,
 			Slug:        slug,
 		})
 
 		if err != nil {
-			return series, nil, FromDBError(err)
+			return nil, nil, FromDBError(err)
 		}
 
-		return series, nil, nil
+		return &series, nil, nil
 	}
 
 	qrs, txn, err := s.database.BeginTx(ctx)
 	if err != nil {
-		return db.Series{}, nil, FromDBError(err)
+		return nil, nil, FromDBError(err)
 	}
 	defer s.database.FinalizeTx(ctx, txn, err)
 
 	series, err := qrs.CreateSeries(ctx, db.CreateSeriesParams{
 		Title:       options.Title,
 		Description: options.Description,
+		AuthorID:    options.UserID,
+		LanguageID:  language.ID,
 		Slug:        slug,
 	})
 	if err != nil {
-		return series, nil, FromDBError(err)
+		return nil, nil, FromDBError(err)
 	}
 
-	tags := make([]db.Tag, 0)
-	if tLen > 0 {
-		t, serviceErr := s.findOrCreateTags(ctx, qrs, findOrCreateTagOptions{
-			Length:   tLen,
-			Tags:     options.Tags,
-			SeriesID: series.ID,
-			UserID:   series.AuthorID,
-		})
-		if serviceErr != nil {
-			return db.Series{}, nil, serviceErr
-		}
-		tags = t
+	tags, serviceErr := s.findOrCreateTags(ctx, qrs, findOrCreateTagOptions{
+		Tags:     options.Tags,
+		SeriesID: series.ID,
+		UserID:   series.AuthorID,
+	})
+	if serviceErr != nil {
+		return nil, nil, serviceErr
 	}
 
-	return series, tags, nil
+	return &series, tags, nil
 }
 
 type SeriesRow struct {
@@ -147,6 +170,7 @@ type SeriesRow struct {
 	Description     string
 	PartsCount      int16
 	LecturesCount   int16
+	IsPublished     bool
 	ReviewAvg       int16
 	ReviewCount     int32
 	AuthorID        int32
@@ -173,6 +197,7 @@ type SeriesDto struct {
 	Description string
 	Parts       int16
 	Lectures    int16
+	IsPublished bool
 	ReviewAvg   int16
 	ReviewCount int32
 	Author      SeriesAuthor
@@ -189,7 +214,7 @@ func mapSeriesRowsToDtos(rows []SeriesRow) []SeriesDto {
 
 	for i, row := range rows {
 		var dto SeriesDto
-		idx := i
+		idx := i - len(dtos)
 		if m, ok := dtos[row.ID]; ok {
 			dto = m.dto
 			idx = m.idx
@@ -201,6 +226,7 @@ func mapSeriesRowsToDtos(rows []SeriesRow) []SeriesDto {
 				Description: row.Description,
 				Parts:       row.PartsCount,
 				Lectures:    row.LecturesCount,
+				IsPublished: row.IsPublished,
 				ReviewAvg:   row.ReviewAvg,
 				ReviewCount: row.ReviewCount,
 				Author: SeriesAuthor{
@@ -232,13 +258,14 @@ func mapSeriesRowsToDtos(rows []SeriesRow) []SeriesDto {
 }
 
 type findPaginatedSeriesRowsOptions struct {
-	LanguageID int32
-	Search     string
-	Tag        string
-	Offset     int32
-	Limit      int32
-	SortBy     string
-	Order      string
+	LanguageID  int32
+	IsPublished bool
+	Search      string
+	Tag         string
+	Offset      int32
+	Limit       int32
+	SortBy      string
+	Order       string
 }
 
 func (s *Services) findPaginatedSeriesRows(ctx context.Context, options findPaginatedSeriesRowsOptions) ([]SeriesRow, int64, *ServiceError) {
@@ -254,6 +281,7 @@ func (s *Services) findPaginatedSeriesRows(ctx context.Context, options findPagi
 		"series"."description",
 		"series"."parts_count",
 		"series"."lectures_count",
+		"series"."is_published",
 		"series"."review_avg",
 		"series"."review_count",
 		"series"."author_id",
@@ -264,26 +292,36 @@ func (s *Services) findPaginatedSeriesRows(ctx context.Context, options findPagi
 	LEFT JOIN "users" ON "series"."author_id" = "users"."id"
 	LEFT JOIN "series_tags" ON "series"."id" = "series_tags"."series_id"
 	LEFT JOIN "tags" ON "series_tags"."tag_id" = "tags"."id"
-	WHERE "series"."is_published" = true AND "languages"."slug" = $1
+	WHERE "series"."language_id" = $1
 	`
-	countQuery := `SELECT COUNT(*) FROM "series" WHERE "series"."is_published" = true AND "languages"."slug" = $1 `
+	countQueryWhere := `WHERE "series"."language_id" = $1 `
+	countQueryJoin := ""
 
+	if options.IsPublished {
+		where := `AND "series"."is_published" = true `
+		query += where
+		countQueryWhere += where
+	}
 	if options.Search != "" {
 		where := fmt.Sprintf(`AND "series"."title" ILIKE $%d `, idx+1)
 		query += where
-		countQuery += where
+		countQueryWhere += where
 		args[idx] = "%" + options.Search + "%"
 		idx++
 	}
 	if options.Tag != "" {
 		where := fmt.Sprintf(`AND "tags"."name" = $%d `, idx+1)
 		query += where
-		countQuery += where
+		countQueryJoin += `
+		LEFT JOIN "series_tags" ON "series"."id" = "series_tags"."series_id"
+		LEFT JOIN "tags" ON "series_tags"."tag_id" = "tags"."id" 
+		`
+		countQueryWhere += where
 		args[idx] = options.Tag
 		idx++
 	}
 
-	countQuery += "LIMIT 1;"
+	countQuery := `SELECT COUNT("series"."id") FROM "series" ` + countQueryJoin + countQueryWhere + "LIMIT 1;"
 	countRow := s.database.RawQueryRow(ctx, countQuery, args[:idx+1])
 	var count int64
 	if err := countRow.Scan(&count); err != nil {
@@ -310,6 +348,7 @@ func (s *Services) findPaginatedSeriesRows(ctx context.Context, options findPagi
 			&s.Description,
 			&s.PartsCount,
 			&s.LecturesCount,
+			&s.IsPublished,
 			&s.ReviewAvg,
 			&s.ReviewCount,
 			&s.AuthorID,
@@ -331,6 +370,7 @@ func (s *Services) findPaginatedSeriesRows(ctx context.Context, options findPagi
 
 type FindPaginatedSeriesOptions struct {
 	LanguageSlug string
+	IsPublished  bool
 	Search       string
 	Tag          string
 	Offset       int32
@@ -350,13 +390,14 @@ func (s *Services) FindPaginatedSeries(ctx context.Context, options FindPaginate
 	}
 
 	rows, count, serviceErr := s.findPaginatedSeriesRows(ctx, findPaginatedSeriesRowsOptions{
-		LanguageID: language.ID,
-		Search:     options.Search,
-		Tag:        options.Tag,
-		Offset:     options.Offset,
-		Limit:      options.Limit,
-		SortBy:     options.SortBy,
-		Order:      options.Order,
+		LanguageID:  language.ID,
+		IsPublished: options.IsPublished,
+		Search:      options.Search,
+		Tag:         options.Tag,
+		Offset:      options.Offset,
+		Limit:       options.Limit,
+		SortBy:      options.SortBy,
+		Order:       options.Order,
 	})
 	if serviceErr != nil {
 		log.WarnContext(ctx, "Error getting series", "error", serviceErr)
@@ -366,7 +407,7 @@ func (s *Services) FindPaginatedSeries(ctx context.Context, options FindPaginate
 	return mapSeriesRowsToDtos(rows), count, nil
 }
 
-func mapSingleSeriesRowsToDto(rows []db.FindSeriesBySlugWithJoinsRow) SeriesDto {
+func mapSingleSeriesRowsToDto(rows []db.FindSeriesBySlugAndLanguageIDWithJoinsRow) *SeriesDto {
 	dto := SeriesDto{
 		ID:          rows[0].ID,
 		Title:       rows[0].Title,
@@ -374,6 +415,7 @@ func mapSingleSeriesRowsToDto(rows []db.FindSeriesBySlugWithJoinsRow) SeriesDto 
 		Description: rows[0].Description,
 		Parts:       rows[0].PartsCount,
 		Lectures:    rows[0].LecturesCount,
+		IsPublished: rows[0].IsPublished,
 		ReviewAvg:   rows[0].ReviewAvg,
 		ReviewCount: rows[0].ReviewCount,
 		Author: SeriesAuthor{
@@ -390,17 +432,26 @@ func mapSingleSeriesRowsToDto(rows []db.FindSeriesBySlugWithJoinsRow) SeriesDto 
 		}
 	}
 
-	return dto
+	return &dto
 }
 
-func (s *Services) FindSeriesBySlugWithJoins(ctx context.Context, slug string) (SeriesDto, *ServiceError) {
-	log := s.log.WithGroup("service.series.FindSeriesBySlugWithJoins").With("slug", slug)
+type FindSeriesBySlugsWithJoinsOptions struct {
+	LanguageSlug string
+	SeriesSlug   string
+	IsPublished  bool
+}
+
+func (s *Services) FindSeriesBySlugsWithJoins(ctx context.Context, opts FindSeriesBySlugsWithJoinsOptions) (*SeriesDto, *ServiceError) {
+	log := s.log.WithGroup("service.series.FindSeriesBySlugWithJoins").With("slug", opts.SeriesSlug)
 	log.InfoContext(ctx, "Getting series by slug")
-	rows, err := s.database.FindSeriesBySlugWithJoins(ctx, slug)
+	rows, err := s.database.FindSeriesBySlugAndLanguageIDWithJoins(ctx, db.FindSeriesBySlugAndLanguageIDWithJoinsParams{
+		Slug:        opts.SeriesSlug,
+		IsPublished: opts.IsPublished,
+	})
 
 	if err != nil {
 		log.WarnContext(ctx, "Error getting series by slug", "error", err)
-		return SeriesDto{}, FromDBError(err)
+		return nil, FromDBError(err)
 	}
 
 	log.InfoContext(ctx, "Series found")
@@ -408,22 +459,34 @@ func (s *Services) FindSeriesBySlugWithJoins(ctx context.Context, slug string) (
 }
 
 type UpdateSeriesOptions struct {
-	Slug        string
-	Title       string
-	Description string
+	UserID       int32
+	LanguageSlug string
+	SeriesSlug   string
+	Title        string
+	Description  string
+	Tags         []string
 }
 
-func (s *Services) UpdateSeries(ctx context.Context, options UpdateSeriesOptions) (db.Series, *ServiceError) {
-	log := s.log.WithGroup("service.series.UpdateSeries").With("slug", options.Slug)
+// TODO: add tags update
+func (s *Services) UpdateSeries(ctx context.Context, options UpdateSeriesOptions) (*db.Series, *ServiceError) {
+	log := s.log.WithGroup("service.series.UpdateSeries").With("slug", options.SeriesSlug)
 	log.InfoContext(ctx, "Updating series")
 
-	series, serviceErr := s.database.FindSeriesBySlug(ctx, options.Slug)
+	series, serviceErr := s.FindSeriesBySlugs(ctx, FindSeriesBySlugsOptions{
+		LanguageSlug: options.LanguageSlug,
+		SeriesSlug:   options.SeriesSlug,
+	})
 	if serviceErr != nil {
 		log.Warn("Series not found", "error", serviceErr)
-		return db.Series{}, FromDBError(serviceErr)
+		return nil, FromDBError(serviceErr)
+	}
+	if series.AuthorID != options.UserID {
+		log.Warn("User is not the author of the series")
+		return nil, NewForbiddenError()
+
 	}
 
-	series, err := s.database.UpdateSeries(ctx, db.UpdateSeriesParams{
+	updatedSeries, err := s.database.UpdateSeries(ctx, db.UpdateSeriesParams{
 		ID:          series.ID,
 		Title:       options.Title,
 		Slug:        utils.Slugify(options.Title),
@@ -431,22 +494,39 @@ func (s *Services) UpdateSeries(ctx context.Context, options UpdateSeriesOptions
 	})
 	if err != nil {
 		log.ErrorContext(ctx, "Error updating series", "error", err)
-		return db.Series{}, FromDBError(err)
+		return nil, FromDBError(err)
 	}
 
 	log.InfoContext(ctx, "Series updated")
-	return series, nil
+	return &updatedSeries, nil
 }
 
-// TODO add constraints to prevent deletion of series with parts
-func (s *Services) DeleteSeries(ctx context.Context, slug string) *ServiceError {
-	log := s.log.WithGroup("service.series.DeleteSeries").With("slug", slug)
+type DeleteSeriesOptions struct {
+	UserID       int32
+	LanguageSlug string
+	SeriesSlug   string
+}
+
+func (s *Services) DeleteSeries(ctx context.Context, opts DeleteSeriesOptions) *ServiceError {
+	log := s.log.WithGroup("service.series.DeleteSeries").With("slug", opts.SeriesSlug)
 	log.InfoContext(ctx, "Deleting series")
 
-	series, serviceErr := s.FindSeriesBySlug(ctx, slug)
+	series, serviceErr := s.FindSeriesBySlugs(ctx, FindSeriesBySlugsOptions{
+		LanguageSlug: opts.LanguageSlug,
+		SeriesSlug:   opts.SeriesSlug,
+	})
 	if serviceErr != nil {
 		log.Warn("Series not found", "error", serviceErr)
 		return FromDBError(serviceErr)
+	}
+	if series.AuthorID != opts.UserID {
+		log.Warn("User is not the author of the series")
+		return NewForbiddenError()
+	}
+	if series.PartsCount > 0 {
+		log.Warn("Series has parts")
+		// TODO: update this to be a constraint error
+		return NewValidationError("series has parts")
 	}
 
 	if err := s.database.DeleteSeriesById(ctx, series.ID); err != nil {
@@ -459,18 +539,27 @@ func (s *Services) DeleteSeries(ctx context.Context, slug string) *ServiceError 
 }
 
 type UpdateSeriesIsPublishedOptions struct {
-	Slug        string
-	IsPublished bool
+	UserID       int32
+	LanguageSlug string
+	SeriesSlug   string
+	IsPublished  bool
 }
 
-func (s *Services) UpdateSeriesIsPublished(ctx context.Context, options UpdateSeriesIsPublishedOptions) (db.Series, *ServiceError) {
-	log := s.log.WithGroup("service.series.UpdateSeriesIsPublished").With("slug", options.Slug, "isPublished", options.IsPublished)
+func (s *Services) UpdateSeriesIsPublished(ctx context.Context, options UpdateSeriesIsPublishedOptions) (*db.Series, *ServiceError) {
+	log := s.log.WithGroup("service.series.UpdateSeriesIsPublished").With("slug", options.SeriesSlug, "isPublished", options.IsPublished)
 	log.InfoContext(ctx, "Updating series is published...")
 
-	series, serviceErr := s.database.FindSeriesBySlug(ctx, options.Slug)
+	series, serviceErr := s.FindSeriesBySlugs(ctx, FindSeriesBySlugsOptions{
+		LanguageSlug: options.LanguageSlug,
+		SeriesSlug:   options.SeriesSlug,
+	})
 	if serviceErr != nil {
 		log.Warn("Series not found", "error", serviceErr)
-		return db.Series{}, FromDBError(serviceErr)
+		return nil, FromDBError(serviceErr)
+	}
+	if series.AuthorID != options.UserID {
+		log.Warn("User is not the author of the series")
+		return nil, NewForbiddenError()
 	}
 
 	if series.IsPublished == options.IsPublished {
@@ -480,20 +569,20 @@ func (s *Services) UpdateSeriesIsPublished(ctx context.Context, options UpdateSe
 	if series.PartsCount == 0 && options.IsPublished {
 		errMsg := "Series must have parts to be published"
 		log.Warn("Series has no parts", "error", errMsg)
-		return db.Series{}, NewValidationError(errMsg)
+		return nil, NewValidationError(errMsg)
 	}
 
 	// TODO: add constraints to prevent unpublishing series with students
 
-	series, err := s.database.UpdateSeriesIsPublished(ctx, db.UpdateSeriesIsPublishedParams{
+	updateSeries, err := s.database.UpdateSeriesIsPublished(ctx, db.UpdateSeriesIsPublishedParams{
 		ID:          series.ID,
 		IsPublished: options.IsPublished,
 	})
 	if err != nil {
 		log.ErrorContext(ctx, "Error publishing series", "error", err)
-		return db.Series{}, FromDBError(err)
+		return nil, FromDBError(err)
 	}
 
 	log.InfoContext(ctx, "Series isPublished updated")
-	return series, nil
+	return &updateSeries, nil
 }
